refactor TrainProgressIndicator into two separate components
